+++ src/redux/store.tsx
@@ -1,11 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 import providersReducer from "./providersSlice";
-import usersReducer from "./ClientsSlice";
+import clientsReducer from "./ClientsSlice";
 
 export const store = configureStore({
   reducer: {
     providers: providersReducer,
-    users: usersReducer,
+    clients: clientsReducer,
   },
 });
